Разгружает модуль form.js
- Создано 2 модуля для того чтобы перенести в них некоторую логику из
  модуля с формой.
- Модуль reset-form отвечает за логику сброса формы при успешной
  отправке данных на сервер и за сброс формы при нажатии на кнопку
  "очистить".
- Модуль set-submit-listener отвечает за добавление слушателя отправки
  на форму adForm.
